Add Docker dev environment and documentation
Adds documentation on running the script and adds a `docker-compose.yml` and
`Dockerfile` needed to run the script inside Docker.

This is a little bit of overhead to run this kind of script - that said, this
setup unblocks me on repairing my dev environment for other tickets (for the
reasons outlined in README.md) and removes some of the monkey patching that I
have to juggle without committing to make the script work on my local machine.