Rewrite AWS report script to use AWS CUR over AWS DBR (#6, #13, #5)
Rewrites the AWS reporting script to use the AWS Cost and Usage Reports
over the AWS Detailed Billing Reports, the latter of which have been
deprecated.

Implementing support for CURs required rewriting the script parsing
code, which required replacing the data structures into which report
data was aggregated, which required rewriting the email generation code.
Ideally, it would have been possible to keep changes as minimal as
possible. Unfortunately, that was not the case.

The intent of implementing support for CURs was to enable email
generation on the first day of the month. (#5)

This rewrite introduces several new features and tweaks:
* Drastically improve report parsing time.
* Visibility over all reported AWS services, not just those that are
  manually configured in the script.
* Fixed a bug where sometimes a diff of $0 would be shown as a negative
  value. (#13)
* The first reported generated in a billing period will no longer show
  N/A as a diff.
* Data for all accounts in a CUR are listed in the generated report,
  even if they are not manually configured in the script.
* Script now uses system temporary directories instead of choosing its
  hardcoded one.
* Using spaces instead of tabs now.

It also introduces some quirks:
* Not all services are tracked in the current database schema, so some
  services will permanently have an "N/A" diff until the schema is
  updated (or the database is replaced altogether).